requirements.txt
fastapi==0.94.0
uvicorn==0.21.0
pymongo==4.3.3
python-dotenv==1.0.0
#selenium==4.8.2
#webdriver-manager==3.8.5
pandas==1.5.3
